+++ src/main/java/ru/urfu/springskblab/service/PizzaService.java
@@ -3,18 +3,20 @@ package ru.urfu.springskblab.service;
 import org.springframework.stereotype.Service;
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @Service
 public class PizzaService implements ProductService {
 
     @PostConstruct
     public void postConstruct() {
-        System.out.println("PizzaService is created");
+        log.info("PizzaService is created");
     }
 
     @PreDestroy
     public void preDestroy() {
-        System.out.println("PizzaService is destroyed");
+        log.info("PizzaService is destroyed");
     }
 
     @Override
